+++ core/metrics/aws/aws.go
@@ -1,4 +1,4 @@
-package metrics
+package aws
 
 import (
 	"context"
