+++ core/metrics/aws/aws_test.go
@@ -1,4 +1,4 @@
-package metrics
+package aws
 
 import (
 	"testing"
