+++ core/metrics/gcp/gcp.go
@@ -1,4 +1,4 @@
-package metrics
+package gcp
 
 import (
 	"context"
