+++ core/metrics/gcp/gcp_test.go
@@ -1,4 +1,4 @@
-package metrics
+package gcp
 
 import (
 	"testing"
