+++ core/metrics/provider.go
@@ -3,14 +3,17 @@ package metrics
 import (
 	"math/rand"
 	"time"
+
+	"github.com/reliablyhq/cli/core/metrics/aws"
+	"github.com/reliablyhq/cli/core/metrics/gcp"
 )
 
 var r = rand.New(rand.NewSource(time.Now().Unix()))
 
 var ProviderFactories = map[string]ProviderFactory{
 	"dummy": func() (Provider, error) { return &DummyProvider{}, nil },
-	"aws":   func() (Provider, error) { return NewAwsCloudWatch() },
-	"gcp":   func() (Provider, error) { return NewGCP() },
+	"aws":   func() (Provider, error) { return aws.NewAwsCloudWatch() },
+	"gcp":   func() (Provider, error) { return gcp.NewGCP() },
 }
 
 type (
@@ -21,13 +24,3 @@ type (
 		GetErrorPercentageMetricForResource(resourceID string, from, to time.Time) (float64, error)
 	}
 )
-
-type DummyProvider struct{}
-
-func (d *DummyProvider) Get99PercentLatencyMetricForResource(resourceID string, from, to time.Time) (float64, error) {
-	return 500 * r.Float64(), nil
-}
-
-func (d *DummyProvider) GetErrorPercentageMetricForResource(resourceID string, from, to time.Time) (float64, error) {
-	return 25 * r.Float64(), nil
-}
